# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Fasal Agriculture Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sensor-data:
    post:
      operationId: submitSensorData
      tags: [sensor]
      summary: Submit IoT sensor data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorDataInput"
      responses:
        "201":
          description: Sensor data stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorData"
    get:
      operationId: getLatestSensorData
      tags: [sensor]
      summary: Get latest sensor data
      responses:
        "200":
          description: Latest sensor reading
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorData"
  /sensor-data/history:
    get:
      operationId: getSensorHistory
      tags: [sensor]
      summary: Get sensor data history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Sensor history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorData"
...
  /insurance/claims:
    get:
      operationId: getInsuranceClaims
      tags: [insurance]
      summary: Get insurance claims
      responses:
        "200":
          description: List of claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsuranceClaim"
    post:
      operationId: createInsuranceClaim
      tags: [insurance]
      summary: Submit an insurance claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInsuranceClaimInput"
      responses:
        "201":
          description: Claim submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsuranceClaim"
...
  /market/listings:
    get:
      operationId: getMarketListings
      tags: [market]
      summary: Get P2P marketplace listings
      responses:
        "200":
          description: Marketplace listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketListing"
    post:
      operationId: createMarketListing
      tags: [market]
      summary: Create a P2P listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarketListingInput"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListing"
  /market/listings/{id}/buy:
    post:
      operationId: buyMarketListing
      tags: [market]
      summary: Buy a marketplace listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListing"
...
  /rewards/connect-wallet:
    post:
      operationId: connectWallet
      tags: [rewards]
      summary: Connect Flow blockchain wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectWalletInput"
      responses:
        "200":
          description: Wallet connected and reward info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardsInfo"
  /rewards/add:
    post:
      operationId: addRewardPoints
      tags: [rewards]
      summary: Add reward points for an activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRewardInput"
      responses:
        "200":
          description: Updated rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardsInfo"
...
components:
  schemas:
...
    SensorDataInput:
      type: object
      properties:
        nitrogen:
          type: number
        phosphorus:
          type: number
        potassium:
          type: number
        ph:
          type: number
        moisture:
          type: number
        deviceId:
          type: ["string", "null"]
      required:
        - nitrogen
        - phosphorus
        - potassium
        - ph
        - moisture
    SensorData:
      type: object
      properties:
        id:
          type: integer
        nitrogen:
          type: number
        phosphorus:
          type: number
        potassium:
          type: number
        ph:
          type: number
        moisture:
          type: number
        deviceId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nitrogen
        - phosphorus
        - potassium
        - ph
        - moisture
        - createdAt
...
    InsuranceClaim:
      type: object
      properties:
        id:
          type: integer
        claimType:
          type: string
        riskLevel:
          type: string
        description:
          type: string
        status:
          type: string
        rewardPoints:
          type: integer
        walletAddress:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - claimType
        - riskLevel
        - description
        - status
        - rewardPoints
        - createdAt
    CreateInsuranceClaimInput:
      type: object
      properties:
        claimType:
          type: string
        description:
          type: string
        walletAddress:
          type: ["string", "null"]
      required:
        - claimType
        - description
...
    MarketListing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        crop:
          type: string
        price:
          type: number
        quantity:
          type: number
        unit:
          type: string
        sellerName:
          type: string
        sellerWallet:
          type: ["string", "null"]
        location:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - crop
        - price
        - quantity
        - unit
        - sellerName
        - location
        - status
        - createdAt
    CreateMarketListingInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        crop:
          type: string
        price:
          type: number
        quantity:
          type: number
        unit:
          type: string
        sellerName:
          type: string
        sellerWallet:
          type: ["string", "null"]
        location:
          type: string
      required:
        - title
        - description
        - crop
        - price
        - quantity
        - unit
        - sellerName
        - location
...
    RewardsInfo:
      type: object
      properties:
        walletAddress:
          type: ["string", "null"]
        totalPoints:
          type: integer
        badges:
          type: array
          items:
            type: string
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/RewardTransaction"
      required:
        - totalPoints
        - badges
        - transactions
    RewardTransaction:
      type: object
      properties:
        id:
          type: integer
        activity:
          type: string
        points:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - activity
        - points
        - createdAt
    ConnectWalletInput:
      type: object
      properties:
        walletAddress:
          type: string
      required:
        - walletAddress
    AddRewardInput:
      type: object
      properties:
        activity:
          type: string
        points:
          type: integer
        walletAddress:
          type: ["string", "null"]
      required:
        - activity
        - points
...