# lib/api-zod/src/generated/types/addRewardInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fasal Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AddRewardInput {
  activity: string;
  points: number;
  /** @nullable */
  walletAddress?: string | null;
}