# lib/api-zod/src/generated/types/marketListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fasal Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketListing {
  id: number;
  title: string;
  description: string;
  crop: string;
  price: number;
  quantity: number;
  unit: string;
  sellerName: string;
  /** @nullable */
  sellerWallet?: string | null;
  location: string;
  status: string;
  createdAt: Date;
}