# lib/api-zod/src/generated/types/rewardTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fasal Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RewardTransaction {
  id: number;
  activity: string;
  points: number;
  createdAt: Date;
}