# lib/api-zod/src/generated/types/rewardsInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fasal Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RewardTransaction } from "./rewardTransaction";

export interface RewardsInfo {
  /** @nullable */
  walletAddress?: string | null;
  totalPoints: number;
  badges: string[];
  transactions: RewardTransaction[];
}