# lib/api-zod/src/generated/types/sensorDataInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fasal Agriculture Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SensorDataInput {
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  ph: number;
  moisture: number;
  /** @nullable */
  deviceId?: string | null;
}